Improve mimic: revert const_set, improve shared tests
